Spell debugging Weapon zone fix Lot of fixes Add gameplay generation Add d2o/d2i files

+++ trunk/Server/Stump.Server.WorldServer/Game/Effects/Handlers/Spells/Buffs/Punishment.cs
@@ -38,7 +38,7 @@ namespace Stump.Server.WorldServer.Game.Effects.Handlers.Spells.Buffs
             if (currentBonus >= limit)
                 return;
 
-            var bonus = (short)token;
+            var bonus = (short)(int)token;
 
             if (bonus + currentBonus > limit)
                 bonus = (short) (limit - currentBonus);
